% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/connect.R
\name{connect}
\alias{connect}
\title{Connection to Lumix Camera}
\usage{
connect(ip = "192.168.54.1", timeout = 5)
}
\arguments{
\item{ip}{}

\item{timeout}{}
}
\description{
This function establishes a connection to a Lumix camera over a
 WiFi network.
}
\seealso{
\code{\link{discover}}
}
\author{
Simon Garnier, \email{garnier@njit.edu}
}
